+++ R/shiny.R
@@ -162,7 +162,7 @@ renderDataTable = function(
       uiFunc = dataTableOutput,
       renderFunc = function(shinysession, name, ...) {
         domain = tempVarsPromiseDomain(outputInfoEnv, outputName = name, session = shinysession)
-
+        removeTimestampFromSnapshot(name)
         promises::with_promise_domain(domain, renderFunc())
       },
       cacheHint = cacheHint
@@ -172,7 +172,7 @@ renderDataTable = function(
       uiFunc = dataTableOutput,
       renderFunc = function(shinysession, name, ...) {
         domain = tempVarsPromiseDomain(outputInfoEnv, outputName = name, session = shinysession)
-
+        removeTimestampFromSnapshot(name)
         promises::with_promise_domain(domain, renderFunc())
       }
     )
@@ -207,6 +207,13 @@ setAll = function(lst, env) {
     assign(name, val, env)
   })
   invisible()
+}
+
+removeTimestampFromSnapshot = function(name) {
+  shiny::snapshotPreprocessInput(paste0(name, "_state"), function(value) {
+    value$time <- NULL
+    value
+  })
 }
 
 # This promise domain is needed to set/unset temporary variables in
@@ -334,10 +341,11 @@ selectCells = function(proxy, selected, ignore.selectable = FALSE) {
 addRow = function(proxy, data, resetPaging = TRUE) {
   if ((is.matrix(data) || is.data.frame(data)) && nrow(data) != 1)
     stop("'data' must be of only one row")
+  rn <- rownames(data); if (!is.null(rn)) rn <- I(rn)
   # must apply unname() after as.list() because a data.table object
   # can't be really unnamed. The names() attributes will be
   # preserved but with empty strings (see #760).
-  invokeRemote(proxy, 'addRow', list(unname(as.list(data)), I(rownames(data)), resetPaging))
+  invokeRemote(proxy, 'addRow', list(unname(as.list(data)), rn, resetPaging))
 }
 
 #' @rdname proxy
@@ -456,11 +464,20 @@ reloadData = function(
 #'   enabled column filters, you should also make sure the attributes of every
 #'   column remain the same, e.g. factor columns should have the same or fewer
 #'   levels, and numeric columns should have the same or smaller range,
-#'   otherwise the filters may never be able to reach certain rows in the data.
+#'   otherwise the filters may never be able to reach certain rows in the data,
+#'   unless you explicitly update the filters with \code{updateFilters()}.
 #' @export
 replaceData = function(proxy, data, ..., resetPaging = TRUE, clearSelection = 'all') {
   dataTableAjax(proxy$session, data, ..., outputId = proxy$rawId)
   reloadData(proxy, resetPaging, clearSelection)
+}
+
+#' @rdname replaceData
+#' @export
+updateFilters = function(proxy, data) {
+  # make sure JS gets an array, not an object
+  filters = unname(columnFilters(data))
+  invokeRemote(proxy, 'updateFilters', list(filters))
 }
 
 invokeRemote = function(proxy, method, args = list()) {
@@ -543,14 +560,15 @@ dataTableAjax = function(session, data, rownames, filter = dataTablesFilter, out
 
 sessionDataURL = function(session, data, id, filter) {
 
-  URLdecode = shinyFun('URLdecode')
   toJSON = shinyFun('toJSON')
   httpResponse = shinyFun('httpResponse')
 
   filterFun = function(data, req) {
     # DataTables requests were sent via POST
-    params = URLdecode(rawToChar(req$rook.input$read()))
+    params = rawToChar(req$rook.input$read())
+    # I don't think the browser would send out nonASCII strings, but keep it as it is
     Encoding(params) = 'UTF-8'
+    # shiny::parseQueryString() calls httpuv::decodeURIComponent() internally and will handle encoding correctly
     params = shiny::parseQueryString(params, nested = TRUE)
 
     res = tryCatch(filter(data, params), error = function(e) {
@@ -569,7 +587,6 @@ sessionDataURL = function(session, data, id, filter) {
 dataTablesFilter = function(data, params) {
   n = nrow(data)
   q = params
-  ci = q$search[['caseInsensitive']] == 'true'
   # users may be updating the table too frequently
   if (length(q$columns) != ncol(data)) return(list(
     draw = as.integer(q$draw),
@@ -580,29 +597,22 @@ dataTablesFilter = function(data, params) {
     DT_rows_current = list()
   ))
 
-  # global searching
-  # for some reason, q$search might be NULL, leading to error `if (logical(0))`
-  if (length(v <- q$search[['value']]) > 0) {
-    if (!identical(q$search[['smart']], 'false')) {
-      v = unlist(strsplit(gsub('^\\s+|\\s+$', '', v), '\\s+'))
-    }
+  # which columns are searchable?
+  searchable = logical(ncol(data))
+  for (j in seq_len(ncol(data))) {
+    if (q$columns[[j]][['searchable']] == 'true') searchable[j] = TRUE
   }
-  if (length(v) == 0) v = ''
-  m = if ((nv <- length(v)) > 1) array(FALSE, c(dim(data), nv)) else logical(n)
-  # TODO: this searching method may not be efficient and need optimization
-  i = if (!identical(v, '')) {
-    for (j in seq_len(ncol(data))) {
-      if (q$columns[[j]][['searchable']] != 'true') next
-      for (k in seq_len(nv)) {
-        i0 = grep2(
-          v[k], as.character(data[, j]), fixed = q$search[['regex']] == 'false',
-          ignore.case = ci
-        )
-        if (nv > 1) m[i0, j, k] = TRUE else m[i0] = TRUE
-      }
-    }
-    which(if (nv > 1) apply(m, 1, function(z) all(colSums(z) > 0)) else m)
-  } else seq_len(n)
+
+  # global searching options (column search shares caseInsensitive)
+  # for some reason, q$search might be NULL, leading to error `if (logical(0))`
+  global_opts = list(
+    smart = !identical(q$search[['smart']], 'false'),
+    regex = q$search[['regex']] != 'false',
+    caseInsensitive = q$search[['caseInsensitive']] == 'true'
+  )
+
+  # start searching with all rows
+  i = seq_len(n)
 
   # search by columns
   if (length(i)) for (j in names(q$columns)) {
@@ -610,21 +620,22 @@ dataTablesFilter = function(data, params) {
     # if the j-th column is not searchable or the search string is "", skip it
     if (col[['searchable']] != 'true') next
     if ((k <- col[['search']][['value']]) == '') next
+    column_opts = list(
+      regex = col[['search']][['regex']] != 'false',
+      caseInsensitive = global_opts$caseInsensitive
+    )
     j = as.integer(j)
-    dj = data[, j + 1]
-    ij = if (is.numeric(dj) || is.Date(dj)) {
-      which(filterRange(dj, k))
-    } else if (is.factor(dj)) {
-      which(dj %in% fromJSON(k))
-    } else if (is.logical(dj)) {
-      which(dj %in% as.logical(fromJSON(k)))
-    } else {
-      grep2(k, as.character(dj), fixed = col[['search']][['regex']] == 'false',
-            ignore.case = ci)
-    }
-    i = intersect(ij, i)
+    dj = data[i, j + 1]
+    i = i[doColumnSearch(dj, k, options = column_opts)]
     if (length(i) == 0) break
   }
+
+  # global searching
+  if (length(i) && any((k <- q$search[['value']]) != '')) {
+    dg = data[i, searchable, drop = FALSE]
+    i = i[doGlobalSearch(dg, k, options = global_opts)]
+  }
+
   if (length(i) != n) data = data[i, , drop = FALSE]
   iAll = i  # row indices of filtered data
 
@@ -680,6 +691,86 @@ dataTablesFilter = function(data, params) {
     DT_rows_all = iAll,
     DT_rows_current = iCurrent
   )
+}
+
+#' Server-side searching
+#'
+#' \code{doGlobalSearch()} can be used to search a data frame given the search
+#' string typed by the user into the global search box of a
+#' \code{\link{datatable}}. \code{doColumnSearch()} does the same for a vector
+#' given the search string typed into a column filter. These functions are used
+#' internally by the default \code{filter} function passed to
+#' \code{\link{dataTableAjax}()}, allowing you to replicate the search results
+#' that server-side processing returns.
+#'
+#' @param x a vector, the type of which determines the expected
+#'   \code{search_string} format
+#' @param search_string a string that determines what to search for. The format
+#'   depends on the type of input, matching what a user would type in the
+#'   associated filter control.
+#' @param options a list of options used to control how searching character
+#'   values works. Supported options are \code{regex}, \code{caseInsensitive}
+#'   and (for global search)
+#'   \href{https://datatables.net/reference/option/search.smart}{\code{smart}}.
+#' @param data a data frame
+#'
+#' @return An integer vector of filtered row indices
+#'
+#' @seealso The column filters section online for search string formats:
+#'   \url{https://rstudio.github.io/DT/}
+#' @seealso Accessing the search strings typed by a user in a Shiny app:
+#'   \url{https://rstudio.github.io/DT/shiny.html}
+#'
+#' @examples
+#' doGlobalSearch(iris, 'versi')
+#' doGlobalSearch(iris, "v.r.i", options = list(regex = TRUE))
+#'
+#' doColumnSearch(iris$Species, '["versicolor"]')
+#' doColumnSearch(iris$Sepal.Length, '4 ... 5')
+#' @export
+doColumnSearch = function(x, search_string, options = list()) {
+  if (length(search_string) == 0 || search_string == '') return(seq_along(x))
+  if (is.numeric(x) || is.Date(x)) {
+    which(filterRange(x, search_string))
+  } else if (is.factor(x)) {
+    which(x %in% fromJSON(search_string))
+  } else if (is.logical(x)) {
+    which(x %in% as.logical(fromJSON(search_string)))
+  } else {
+    grep2(
+      search_string, as.character(x),
+      fixed = !(options$regex %||% FALSE),
+      ignore.case = options$caseInsensitive %||% TRUE
+    )
+  }
+}
+
+#' @rdname doColumnSearch
+#' @export
+doGlobalSearch = function(data, search_string, options = list()) {
+  n = nrow(data)
+  if (length(v <- search_string) > 0) {
+    if (options$smart %||% TRUE) {
+      # https://datatables.net/reference/option/search.smart
+      v = unlist(strsplit(gsub('^\\s+|\\s+$', '', v), '\\s+'))
+    }
+  }
+  if (length(v) == 0) v = ''
+  m = if ((nv <- length(v)) > 1) array(FALSE, c(dim(data), nv)) else logical(n)
+  # TODO: this searching method may not be efficient and need optimization
+  if (!identical(v, '')) {
+    for (j in seq_len(ncol(data))) {
+      for (k in seq_len(nv)) {
+        i0 = grep2(
+          v[k], as.character(data[, j]),
+          fixed = !(options$regex %||% FALSE),
+          ignore.case = options$caseInsensitive %||% TRUE
+        )
+        if (nv > 1) m[i0, j, k] = TRUE else m[i0] = TRUE
+      }
+    }
+    which(if (nv > 1) apply(m, 1, function(z) all(colSums(z) > 0)) else m)
+  } else seq_len(n)
 }
 
 # when both ignore.case and fixed are TRUE, we use grep(ignore.case = FALSE,